Updates to Customizing Plots lecture
This lecture uses the Lahman package. It calls an object "Master" which is no longer in the package. The object is now called "People" in the package. I went through the code and updated it to reflect this change. The simplest thing to do was to define "Master" as "People". In the code, rather than updating all the calls to "Master".

Co-Authored-By: Jesse Lecy <1209099+lecy@users.noreply.github.com>