Changed object name
This lecture uses the Lahman package. It calls an object "Master" which is no longer in the package. The object is now called "People" in the package. I just did a find and replace for the terms. That was easier and cleaner than the edits I made to the earlier version.